fix: improve logging format
